Separated TileMapItem (visual data) from TerrainClass (behavioral data)
It seems likely that TerrainClass will end up being set to a different position/scale than TileMapItem.  My thinking is that TileMapItem will be super-imposed on top of TerrainClass with a 1/2 tile displacement, so that internal terrain information for tiles aligns with what the user sees visually.  Benefit is that it should be pretty simple to handle in the code.  Drawback is it might take a little getting used to -- built-in vis tool would be handy!

Known Alternatives are:
 * Underlying Terrain map is 4x larger than visual tile map.
 * Terrain map and tile map both have lots of special data to deal with cornering logic and clipping behavior along terrain transitions.
 * Scrap tile maps entirely and use a polygon-based rendering system (for appearance) with full 3D rigid body physics concepts to control gameplay behavior (<-- this would be my suggestion if we returned to Unity!)