Gameover screen and respawn system (#35)
Stops the character from attacking when you're dead
Stops the character from collecting items
Display a game over message
Remove all enemies and items of the screen when respawning
Separate Entity destruction from being killed (so that they don't all gives items on respawn)
Changed item logic to be active instead of passive: "It is the Player that acts on the Item, not the Item that is aware that a player can pick it up". This will scale better.